requirements.txt
openenv-core[core]>=0.2.0
gradio>=6.0.0
fastmcp>=2.14.5
pydantic>=2.0
mcp>=1.26.0
httpx>=0.27
pandas>=2.0
